Extract TournamentStream component
