+++ app/features/tournament-bracket/routes/to.$id.matches.$mid.tsx
@@ -358,6 +358,7 @@ export const loader = ({ params }: LoaderFunctionArgs) => {
   };
 };
 
+// xxx: when match is locked still show room pass + pool?
 // xxx: rename points -> score to match ingame in the UI?
 // xxx: lacks padding on top as non-admin http://localhost:5800/to/10/matches/997
 export default function TournamentMatchPage() {
